mlops-project/requirements.txt
python-dotenv
prefect
yfinance
pandas
lightgbm
scikit-learn
matplotlib
seaborn
joblib
numpy
